server/requirements.txt
astroid==3.3.5
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.2
certifi==2024.8.30
charset-normalizer==3.4.0
clarabel==0.9.0
click==8.1.7
comm==0.2.2
cvxpy==1.5.3
debugpy==1.8.7
decorator==5.1.1
dill==0.3.9
diskcache==5.6.3
ecos==2.0.14
et-xmlfile==1.1.0
executing==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
frozendict==2.4.6
greenlet==3.1.1
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack==1.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
osqp==0.6.7.post3
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
peewee==3.17.7
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
pynvim==0.5.0
pyportfolioopt==1.5.5
python-dateutil==2.9.0.post0
pytz==2024.2
pyuwsgi==2.0.26
pyzmq==26.2.0
qdldl==0.1.7.post4
requests==2.32.3
scipy==1.14.1
scs==3.2.7
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
uWSGI==2.0.27
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
yfinance==0.2.44
